% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_mass_balance.R
\name{plot.MassBalance}
\alias{plot.MassBalance}
\title{Plot MassBalance object}
\usage{
\method{plot}{MassBalance}(x, ...)
}
\arguments{
\item{stuff}{}
}
\value{
stuff
}
\description{
stuff
}
\details{
stuff
}
\examples{
stuff <- 1
}

