% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RRQP-package.R
\docType{package}
\name{RRQP}
\alias{RRQP}
\alias{RRQP-package}
\title{R Implementation of EA River Quality Planning Tools}
\description{
An R implementation of the Fortran River Quality Planning tools. It
allows you to generate mass balance based estimates of water quality
downstream of a single discharge. Downstream water quality estimates can be
generated using summary statistics or by entering observed data. The R
implementation should be considered more flexible than the Fortran GUI
versions, calibrated can be edited, distributions can be specified and
multiple 'RQPs' can be easily be chained.
}
\references{
stuff
}

