% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary_mass_balance.R
\name{summary.MassBalance}
\alias{summary.MassBalance}
\title{Summary for MassBalance object}
\usage{
\method{summary}{MassBalance}(x, ...)
}
\arguments{
\item{stuff}{}
}
\value{
stuff
}
\description{
stuff
}
\details{
stuff
}
\examples{
stuff <- 1
}

