% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mass_balance.R
\name{SummaryMassBalance}
\alias{SummaryMassBalance}
\title{Simple MC Water Quality Mass Balance using summary statistics}
\usage{
SummaryMassBalance(us.q.m, us.q.sd, us.f.m, us.f.sd, dis.q.m, dis.q.sd, dis.f.m,
  dis.f.sd, shots = 2500, cor = 1L, kSize = 10000, dist = "logNormal")
}
\arguments{
\item{stuff}{}
}
\value{
stuff
}
\description{
\code{SimpleMassBalance} returns a "MassBalance" object.
}
\details{
use \code{SummaryMassBalance} as user input layer to access
         \code{SimpleMassBalance} with user inputting only summary stats for
         us and discharge
}
\examples{
stuff <- 1
}

